Refactor test and talon  groups
