+++ healthcheck/src/main/kotlin/org/strykeforce/thirdcoast/healthcheck/groups/TestGroup.kt
@@ -1,21 +1,18 @@
-package org.strykeforce.thirdcoast.healthcheck
+package org.strykeforce.thirdcoast.healthcheck.groups
 
-import com.ctre.phoenix.motorcontrol.can.TalonSRX
 import kotlinx.html.TagConsumer
 import kotlinx.html.div
 import kotlinx.html.h2
 import kotlinx.html.table
 import mu.KotlinLogging
-import org.strykeforce.thirdcoast.healthcheck.TestGroup.TestGroupState.*
-import org.strykeforce.thirdcoast.healthcheck.tests.TalonPosition
-import org.strykeforce.thirdcoast.healthcheck.tests.TalonPositionTest
-import org.strykeforce.thirdcoast.healthcheck.tests.TalonTimedTest
-
-private val logger = KotlinLogging.logger {}
+import org.strykeforce.thirdcoast.healthcheck.HealthCheck
+import org.strykeforce.thirdcoast.healthcheck.Reportable
+import org.strykeforce.thirdcoast.healthcheck.Test
+import org.strykeforce.thirdcoast.healthcheck.groups.TestGroup.TestGroupState.*
 
 abstract class TestGroup(val healthCheck: HealthCheck) : Test {
     override var name = "name not set"
-
+    private val logger = KotlinLogging.logger {}
     protected val tests = mutableListOf<Test>()
     private var state = STARTING
     private lateinit var iterator: Iterator<Test>
@@ -70,33 +67,4 @@ abstract class TestGroup(val healthCheck: HealthCheck) : Test {
         STOPPED
     }
 
-}
-
-
-class TalonGroup(healthCheck: HealthCheck) : TestGroup(healthCheck) {
-    var talons = emptyList<TalonSRX>()
-
-    @Suppress("unused")
-    fun timedTest(init: TalonTimedTest.() -> Unit): Test {
-        val spinTest = TalonTimedTest(this)
-        spinTest.init()
-        tests.add(spinTest)
-        return spinTest
-    }
-
-    @Suppress("unused")
-    fun positionTest(init: TalonPositionTest.() -> Unit): Test {
-        val positionTest = TalonPositionTest(this)
-        positionTest.init()
-        tests.add(positionTest)
-        return positionTest
-    }
-
-    @Suppress("unused")
-    fun positionTalon(init: TalonPosition.() -> Unit): Test {
-        val position = TalonPosition(this)
-        position.init()
-        tests.add(position)
-        return position
-    }
 }
